neurons/miners/openai/requirements.txt
openai==1.9.0
langchain==0.1.13
python-dotenv
langchain-openai
langchain-community
langchain-anthropic
langchain_anthropic
anthropic==0.21.3
git+https://github.com/AIT-Protocol/NumPAL.git@main#egg=NumPAL